move if to bundle package
